kochi_metro_backend/requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn
joblib
pymoo
prisma[asyncpg]
python-dotenv
imbalanced-learnadd 
colorama
tabulate
aiohttp